+++ python/dgl/__init__.py
@@ -29,7 +29,6 @@ from .heterograph import DGLHeteroGraph as DGLGraph  # pylint: disable=reimporte
 from .traversal import *
 from .transform import *
 from .propagate import *
-from .udf import NodeBatch, EdgeBatch
 
 from ._deprecate.graph import DGLGraph as DGLGraphStale
 from ._deprecate.nodeflow import *
