+++ tests/lint/pylintrc
@@ -389,7 +389,7 @@ ignored-classes=optparse.Values,thread._local,_thread._local
 # (useful for modules/projects where namespaces are manipulated during runtime
 # and thus existing member attributes cannot be deduced by static analysis. It
 # supports qualified module names, as well as Unix pattern matching.
-ignored-modules=dgl.backend,dgl._api_internal
+ignored-modules=dgl.backend,dgl._api_internal,dgl._deprecate
 
 # Show a hint with possible names when a member name was not found. The aspect
 # of finding the hint is based on edit distance.
